Postprocess mask opening after mesh rendering.
This is because, in the original mesh rendering code, all the depth edges are marked invalid even though it is not deoccluded.
There areas are thin, so apply mask opening to remove them.
Note that this discards slim masks around the image borders, so we add border padding before opening.